Training and predicting uv components
