+++ api/app/weather_models/regression_model.py
@@ -1,14 +1,12 @@
 import logging
 import math
-from datetime import datetime
-from sklearn.exceptions import NotFittedError
 from sklearn.linear_model import LinearRegression
 from typing import Dict, List, Protocol
 from collections import defaultdict
 from abc import abstractmethod
 from app.db.models.observations import HourlyActual
 from app.db.models.weather_models import ModelRunPrediction
-from app.weather_models import construct_interpolated_noon_prediction
+from app.weather_models.linear_model import LinearModel
 from app.weather_models.sample import Samples
 
 logger = logging.getLogger(__name__)
@@ -34,33 +32,6 @@ class RegressionModelProto(Protocol):
 
     @abstractmethod
     def predict(self, hour: int, model_wind_dir: List[List[int]]): raise NotImplementedError
-
-
-class LinearModel():
-    _models: defaultdict[int, LinearRegression]
-    _samples: Samples
-
-    def __init__(self):
-        self._models = defaultdict(LinearRegression)
-        self._samples = Samples()
-
-    def append_x(self, value, timestamp: datetime):
-        self._samples.append_x(value, timestamp)
-
-    def append_y(self, value, timestamp: datetime):
-        self._samples.append_y(value, timestamp)
-
-    def train(self):
-        for hour in self._samples.hours():
-            self._models[hour].fit(self._samples.np_x(hour), self._samples.np_y(hour))
-
-    def predict(self, hour: int, model_wind_dir: List[List[int]]):
-        try:
-            prediction = self._models[hour].predict(model_wind_dir)
-            logger.info("Predicted wind dir for model: %s, hour: %s, prediction: %s", self._key, hour, prediction)
-            return prediction[0]
-        except NotFittedError as _:
-            return None
 
 
 class RegressionModel(RegressionModelProto):
@@ -100,46 +71,3 @@ class RegressionModel(RegressionModelProto):
             # Using two variables, the interpolated temperature value, and the hour of the day.
             self._linear_model.append_x(model_value, actual.weather_date)
             self._linear_model.append_y(actual_value, actual.weather_date)
-
-
-class RegressionModelsV2:
-    """ Class for storing regression models.
-    TODO: migrate other models to this once wind direction is verified
-    """
-
-    def __init__(self):
-        self._model_keys: List[str] = list(model_2_actual_keys.keys())
-        self._models: List[RegressionModelProto] = [
-            RegressionModel(model_key=ModelRunPrediction.wdir_tgl_10.name)
-        ]
-
-    def add_samples(self, prediction: ModelRunPrediction, actual: HourlyActual):
-        for model in self._models:
-            model.add_sample(prediction, actual)
-
-    def collect_data(self, query):
-        # We need to keep track of previous so that we can do interpolation for the global model.
-        prev_actual = None
-        prev_prediction = None
-        for actual, prediction in query:
-            if prev_actual != actual and prediction is not None:
-                if (prev_actual is not None
-                        and prev_prediction is not None
-                        and prev_actual.weather_date.hour == 20
-                        and prediction.prediction_timestamp.hour == 21
-                        and prev_prediction.prediction_timestamp.hour == 18):
-                    # If there's a gap in the data (like with the GLOBAL model) - then make up
-                    # a noon prediction using interpolation, and add it as a sample.
-                    noon_prediction = construct_interpolated_noon_prediction(
-                        prev_prediction, prediction, self._model_keys)
-                    self.add_samples(noon_prediction, prev_actual)
-
-                self.add_samples(prediction, actual)
-                prev_prediction = prediction
-            prev_actual = actual
-
-    def train(self):
-        # iterate through the data, creating a regression model for each variable
-        # and each hour.
-        for model in self._models:
-            model.train()
